+++ web/WEB-INF/views/material/materialList.jsp
@@ -59,7 +59,7 @@
                 </div>
                 <div class="f-right text-right">
                     <button type="button" id="manager_input" class="btn btn-primary"><i class="fa fa-pencil" aria-hidden="true"></i> 등록</button>
-                    <button type="button" id="account_excel_download" class="btn btn-success"><i class="fa fa-file-excel-o" aria-hidden="true"></i> Excel <button>
+                    <button type="button" id="account_excel_download" class="btn btn-success"><i class="fa fa-file-excel-o" aria-hidden="true"></i> Excel </button>
                     <button type="button" id="manager_delete" class="btn btn-danger"><i class="fa fa-trash" aria-hidden="true"></i> 삭제 </button>
                 </div>
             </div>
